requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-appconf==1.0.5
django-compressor==3.1
django-cors-headers==3.10.1
django-libsass==0.9
libsass==0.21.0
mysqlclient==2.1.0
#mysqlclient==2.0.1
unipath==1.1
#six==1.15.0

#rcssmin==1.0.6
#rjsmin==1.1.0
